requirements.txt
Flask==3.0.3
requests==2.32.3
pandas==2.2.2
apscheduler==3.10.4
alpaca-py==0.17.0
pytz==2024.1